+++ src/main/java/com/sparta/spring_deep/_delivery/config/WebSecurityConfig.java
@@ -2,9 +2,8 @@ package com.sparta.spring_deep._delivery.config;
 
 import com.sparta.spring_deep._delivery.config.security.JwtAuthenticationFilter;
 import com.sparta.spring_deep._delivery.config.security.JwtAuthorizationFilter;
-import com.sparta.spring_deep._delivery.domain.user.UserDetailsServiceImpl;
+import com.sparta.spring_deep._delivery.domain.user.details.UserDetailsServiceImpl;
 import com.sparta.spring_deep._delivery.util.JwtUtil;
-import jakarta.servlet.Filter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
 import org.springframework.context.annotation.Bean;
