+++ src/main/java/com/sparta/spring_deep/_delivery/config/security/JwtAuthorizationFilter.java
@@ -1,6 +1,6 @@
 package com.sparta.spring_deep._delivery.config.security;
 
-import com.sparta.spring_deep._delivery.domain.user.UserDetailsServiceImpl;
+import com.sparta.spring_deep._delivery.domain.user.details.UserDetailsServiceImpl;
 import com.sparta.spring_deep._delivery.util.JwtUtil;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.FilterChain;
