+++ src/main/java/com/sparta/spring_deep/_delivery/domain/menu/Menu.java
@@ -1,7 +1,7 @@
 package com.sparta.spring_deep._delivery.domain.menu;
 
 import com.sparta.spring_deep._delivery.common.BaseEntity;
-import com.sparta.spring_deep._delivery.domain.user.User;
+import com.sparta.spring_deep._delivery.domain.user.entity.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
