+++ src/main/java/com/sparta/spring_deep/_delivery/domain/menu/MenuService.java
@@ -3,8 +3,8 @@ package com.sparta.spring_deep._delivery.domain.menu;
 import com.sparta.spring_deep._delivery.domain.ai.Ai;
 import com.sparta.spring_deep._delivery.admin.repository.AiRepository;
 import com.sparta.spring_deep._delivery.domain.ai.GoogleAiService;
-import com.sparta.spring_deep._delivery.domain.user.User;
-import com.sparta.spring_deep._delivery.domain.user.UserRepository;
+import com.sparta.spring_deep._delivery.domain.user.entity.User;
+import com.sparta.spring_deep._delivery.domain.user.repository.UserRepository;
 import jakarta.persistence.EntityNotFoundException;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
