+++ src/main/java/com/sparta/spring_deep/_delivery/domain/order/Order.java
@@ -3,7 +3,7 @@ package com.sparta.spring_deep._delivery.domain.order;
 import com.sparta.spring_deep._delivery.common.BaseEntity;
 import com.sparta.spring_deep._delivery.domain.address.Address;
 import com.sparta.spring_deep._delivery.domain.restaurant.entity.Restaurant;
-import com.sparta.spring_deep._delivery.domain.user.User;
+import com.sparta.spring_deep._delivery.domain.user.entity.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
