--- src/main/java/com/sparta/spring_deep/_delivery/domain/user/IsPublic.java
@@ -1,6 +1,0 @@
-package com.sparta.spring_deep._delivery.domain.user;
-
-public enum IsPublic {
-    PUBLIC,
-    PRIVATE
-}
