--- src/main/java/com/sparta/spring_deep/_delivery/domain/user/UserControllerAdvice.java
@@ -1,11 +1,0 @@
-package com.sparta.spring_deep._delivery.domain.user;
-
-import org.springframework.web.bind.annotation.ControllerAdvice;
-
-@ControllerAdvice
-public class UserControllerAdvice {
-
-    private UserService userService;
-
-
-}
