+++ src/main/java/com/sparta/spring_deep/_delivery/domain/user/entity/IsPublic.java
@@ -1,0 +1,6 @@
+package com.sparta.spring_deep._delivery.domain.user.entity;
+
+public enum IsPublic {
+    PUBLIC,
+    PRIVATE
+}
