+++ src/main/java/com/sparta/spring_deep/_delivery/domain/user/entity/UserRole.java
@@ -1,4 +1,4 @@
-package com.sparta.spring_deep._delivery.domain.user;
+package com.sparta.spring_deep._delivery.domain.user.entity;
 
 public enum UserRole {
     CUSTOMER,
