+++ src/main/java/com/sparta/spring_deep/_delivery/domain/user/service/UserService.java
@@ -1,13 +1,17 @@
-package com.sparta.spring_deep._delivery.domain.user;
+package com.sparta.spring_deep._delivery.domain.user.service;
 
-import com.sparta.spring_deep._delivery.domain.auth.LoginRequestDto;
-import com.sparta.spring_deep._delivery.domain.auth.LoginResponseDto;
+import com.sparta.spring_deep._delivery.domain.user.details.UserDetailsImpl;
+import com.sparta.spring_deep._delivery.domain.user.dto.LoginRequestDto;
+import com.sparta.spring_deep._delivery.domain.user.dto.LoginResponseDto;
+import com.sparta.spring_deep._delivery.domain.user.dto.PasswordChangeDto;
+import com.sparta.spring_deep._delivery.domain.user.dto.UserDto;
+import com.sparta.spring_deep._delivery.domain.user.entity.User;
+import com.sparta.spring_deep._delivery.domain.user.repository.UserRepository;
 import com.sparta.spring_deep._delivery.util.JwtUtil;
 import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.apache.catalina.Authenticator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
